Add AbstractConfigManager
